+++ docs/changelog.rst
@@ -14,6 +14,7 @@ Added:
   ``${env:VARIABLE}``.
 * The ``-b``, ``--basedir`` argument to override the base directory for storage. In
   contrast to ``--temp-basedir`` the directory is not deleted later.
+* ``:rename`` and ``mark-rename`` commands to rename files starting from a common base.
 
 Fixed:
 ^^^^^^